refactor: folder, set, user, other
