--- client/src/modules/dashboard/DashboardPage.tsx
@@ -1,104 +1,0 @@
-import React from 'react';
-import type { NextPage } from 'next';
-import { useRouter } from 'next/router';
-import { useQuery } from 'react-query';
-import { setApi } from 'apis/setApi';
-import { folderApi } from 'apis/folderApi';
-import { authApi } from 'apis/authApi';
-import { useUserStore } from 'storage/useUserStore';
-import { CardBox } from 'ui/CardBox';
-import { Button } from 'ui/Button';
-import { FolderForm } from 'modules/FolderForm';
-import { FolderInterface, SetInterface } from 'interfaces';
-import style from 'styles/pages/home.module.scss';
-import { ActionList } from '@src/ui/ActionList';
-
-type TabsOptions = 'sets' | 'folders';
-type TabsProps = { onTabSelect: React.Dispatch<React.SetStateAction<TabsOptions>>; selectedTab: TabsOptions };
-type StateType = { userSets: SetInterface[]; userFolders: FolderInterface[] };
-
-const DashboardPage: NextPage = () => {
-  const [selectedTab, setSelectedTab] = React.useState<TabsOptions>('sets');
-
-  return (
-    <div className={style.container}>
-      <DisplayData selectedTab={selectedTab} />
-    </div>
-  );
-};
-
-const DisplayData: React.FC<{ selectedTab: TabsOptions }> = ({ selectedTab }) => {
-  const router = useRouter();
-  const { signAccess } = useUserStore();
-  const [shownFolder, setShownFolder] = React.useState(false);
-  const toggleShownFolder = () => setShownFolder(!shownFolder);
-
-  const user = useQuery('user', () => authApi.me(signAccess)); // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-  const userSets = useQuery('userSets', () => setApi.getByUser(user.data!.id), { enabled: !!user.data }); // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-  const userFolders = useQuery('userFolders', () => folderApi.getByUser(user.data!.id), { enabled: !!user.data });
-  const set = useQuery('sets', () => setApi.get(user.data?.id), { enabled: !!user.data || user.isFetched });
-  const folder = useQuery('folders', () => folderApi.get(user.data?.id), { enabled: !!user.data || user.isFetched });
-
-  const [state, setState] = React.useState<StateType>({ userSets: [], userFolders: [] });
-  React.useEffect(() => {
-    if (userSets.data) setState((prev) => ({ ...prev, userSets: userSets.data }));
-    if (userFolders.data) setState((prev) => ({ ...prev, userFolders: userFolders.data }));
-  }, [userSets.data, userFolders.data]);
-
-  return (
-    <>
-      {/* {selectedTab === 'sets' ? (
-        <>
-          <section>
-            <div className={style.cardlist}>
-              {state.userSets.length ? (
-                state.userSets.map((content) => <CardBox key={content.id} content={content.title} id={content.id} type="set" />)
-              ) : (
-                <div className={style.emptyCard}>
-                  <p>You don&#39;t have any study sets yet</p>
-                  <Button onClick={() => router.push('/create-set')}>Create your first study set</Button>
-                </div>
-              )}
-            </div>
-          </section>
-          <section>
-            <h2>{'Discover solutions from other users'}</h2>
-            <div className={style.cardlist}>
-              {set.data?.length
-                ? set.data.map((content) => <CardBox key={content.id} content={content.title} id={content.id} type="set" />)
-                : undefined}
-            </div>
-          </section>
-        </>
-      ) : undefined} */}
-
-      <section>
-        <ActionList
-          header={{ title: 'Folders' }}
-          data={state.userFolders}
-          keyExtractor={(item) => item.id}
-          renderItem={(item, index) => (
-            <ActionList.Link href={`folder/${item.id}`} isFirst={index === 0}>
-              {item.name}
-            </ActionList.Link>
-          )}
-        />
-        {/* <div className={style.cardlist}>
-          {state.userFolders.length ? (
-            state.userFolders.map((content) => <CardBox key={content.id} content={content.name} id={content.id} type="folder" />)
-          ) : (
-            <>
-              <div className={style.emptyCard}>
-                <p>You don&#39;t have any folders yet</p>
-                <Button onClick={toggleShownFolder}>Create your first folder</Button>
-              </div>
-              <FolderForm isOpen={shownFolder} onClose={toggleShownFolder} />
-            </>
-          )}
-        </div> */}
-      </section>
-    </>
-  );
-};
-
-export default DashboardPage;
