+++ client/src/pages/sets/[id]/index.tsx
@@ -1,0 +1,3 @@
+import SetDetails from '@src/modules/set/SetDetails';
+
+export default SetDetails;
