+++ client/src/pages/sets/[id]/write.tsx
