+++ client/src/pages/sets/index.tsx
@@ -1,0 +1,3 @@
+import SetList from '@src/modules/set/SetList';
+
+export default SetList;
